requirements.txt
appdirs==1.4.3
beautifulsoup4==4.5.3
coverage==4.3.4
Django==1.11.1
django-appconf==1.0.2
django-classy-tags==0.8.0
django-cookie-law==1.0.13
django-coverage==1.2.4
django-debug-toolbar==1.7
django-imagekit==4.0
django-ipware==1.1.6
django-modeladmin-reorder==0.2
django-webtest==1.9.1
flake8==3.3.0
maxminddb==1.3.0
maxminddb-geolite2==2017.404
mccabe==0.6.1
olefile==0.44
packaging==16.8
pilkit==2.0
Pillow==4.0.0
psycopg2==2.7.1
pycodestyle==2.3.1
pyflakes==1.5.0
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
sqlparse==0.2.3
twentytab-sortable==0.11
uWSGI==2.0.14
waitress==1.0.2
WebOb==1.7.2
WebTest==2.0.27
